Moved auth code to folder
