+++ admin/authentication/logoutEndpoint.js
